refactor: cleanup the packages slightly
Move some files around so they fit into packages that are more correct for their purpose.
