+++ src/main/kotlin/com/chesire/orochi/api/kitsu/KitsuEndpoint.kt
@@ -1,0 +1,7 @@
+package com.chesire.orochi.api.kitsu
+
+object KitsuEndpoint {
+    object OAuth {
+        const val Token = "https://kitsu.io/api/oauth/token"
+    }
+}
