--- src/main/kotlin/com/chesire/orochi/api/kitsu/model/KitsuEndpoint.kt
@@ -1,2 +1,0 @@
-package com.chesire.orochi.api.kitsu.model
-
