+++ src/main/kotlin/com/chesire/orochi/plugins/httpclient/HttpClient.kt
@@ -1,6 +1,5 @@
 package com.chesire.orochi.plugins.httpclient
 
-import io.ktor.application.Application
 import io.ktor.client.HttpClient
 import io.ktor.client.features.json.JsonFeature
 import io.ktor.client.features.logging.DEFAULT
@@ -8,7 +7,7 @@ import io.ktor.client.features.logging.LogLevel
 import io.ktor.client.features.logging.Logger
 import io.ktor.client.features.logging.Logging
 
-fun Application.configureHttpClient(): HttpClient {
+fun configureHttpClient(): HttpClient {
     return HttpClient() {
         install(JsonFeature)
         install(Logging) {
